pyproject.toml
[project]
# https://pip.pypa.io/en/stable/reference/build-system/pyproject-toml/
name = 'questdb-connect'
#version = '1.0.10' # Standalone production version (with engine)
version = '0.0.101' # testing version
authors = [{ name = 'questdb.io', email = 'miguel@questdb.io' }]
description = "SqlAlchemy library"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    'Intended Audience :: Developers',
    'Topic :: Database',
    'License :: OSI Approved :: Apache Software License',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
dependencies = []

[project.urls]
'Homepage' = "https://github.com/questdb/questdb-connect/"
'Bug Tracker' = "https://github.com/questdb/questdb-connect/issues/"
'QuestDB GitHub' = "https://github.com/questdb/questdb/"
'QuestDB Docs' = "https://questdb.io/docs/"

[project.entry-points.'sqlalchemy.dialects']
questdb = 'questdb_connect.dialect:QuestDBDialect'

[project.entry-points.'superset.db_engine_specs']
questdb = 'qdb_superset.db_engine_specs.questdb:QuestDbEngineSpec'

[project.optional-dependencies]
test = [
    'psycopg2-binary~=2.9.6',
    'SQLAlchemy>=1.4, <2',
    'apache-superset>=3.0.0',
    'sqlparse==0.4.4',
    'pytest~=7.3.0',
    'pytest_mock~=3.11.1',
    'black~=23.3.0',
    'ruff~=0.0.269',
]

[tool.ruff]
# https://github.com/charliermarsh/ruff#configuration
select = ["PL", "RUF", "TCH", "TID", "PT", "C4", "B", "S", "I"]
line-length = 120
exclude = [
    ".pytest_cache",
    ".questdb_data",
    ".git",
    ".ruff_cache",
    "venv",
    "dist",
    "questdb_connect.egg-info",
]

[tool.ruff.pylint]
max-branches = 20
max-args = 10

[tool.ruff.per-file-ignores]
'tests/test_dialect.py' = ['S101']
'tests/test_types.py' = ['S101']
'tests/test_superset.py' = ['S101']
'tests/conftest.py' = ['S608']
'src/examples/sqlalchemy_raw.py' = ['S608']
'src/examples/server_utilisation.py' = ['S311']
